Move SourceBuild stuff to its own file
This just helps organize code a little nicer and reduce conflicts for
unrelated things.
